Full refactoring code

+++ src/main/java/pro/sky/finalprojectsky/repository/AdsRepository.java
@@ -1,16 +1,11 @@
 package pro.sky.finalprojectsky.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import pro.sky.finalprojectsky.model.Ads;
-
-
+import pro.sky.finalprojectsky.entity.Ads;
 import java.util.List;
 
-/**
- * Repository AdsRepository (advertisement/объявление).
- */
-public interface AdsRepository extends JpaRepository<Ads, Long> {
+public interface AdsRepository extends JpaRepository<Ads, Integer> {
 
     List<Ads> findAllByAuthorId(long id);
+
 }
